+++ packages/assets-controllers/src/NftDetectionController.ts
@@ -740,7 +740,7 @@ export class NftDetectionController extends BaseController<
             kind,
             image: imageUrl,
             imageSmall: imageThumbnailUrl,
-            metadata: { imageOriginal: imageOriginalUrl } = {},
+            metadata,
             name,
             description,
             attributes,
@@ -751,6 +751,9 @@ export class NftDetectionController extends BaseController<
             collection,
             chainId,
           } = nft.token;
+
+          // Use a fallback if metadata is null
+          const { imageOriginal: imageOriginalUrl } = metadata || {};
 
           let ignored;
           /* istanbul ignore else */