+++ packages/phishing-controller/src/index.ts
@@ -8,4 +8,4 @@ export type {
 } from './PhishingDetector';
 export { PhishingDetector } from './PhishingDetector';
 export type { PhishingDetectionScanResult } from './types';
-export { PhishingDetectorResultType } from './types';
+export { PhishingDetectorResultType, RecommendedAction } from './types';