+++ packages/transaction-controller/src/utils/batch.ts
@@ -1,7 +1,8 @@
 import type EthQuery from '@metamask/eth-query';
 import { rpcErrors } from '@metamask/rpc-errors';
 import type { Hex } from '@metamask/utils';
-import { createModuleLogger } from '@metamask/utils';
+import { bytesToHex, createModuleLogger } from '@metamask/utils';
+import { parse, v4 } from 'uuid';
 
 import {
   doesChainSupportEIP7702,
@@ -9,15 +10,19 @@ import {
   isAccountUpgradedToEIP7702,
 } from './eip7702';
 import {
+  getBatchSizeLimit,
   getEIP7702SupportedChains,
   getEIP7702UpgradeContractAddress,
 } from './feature-flags';
 import { validateBatchRequest } from './validation';
-import type { TransactionController, TransactionControllerMessenger } from '..';
+import type {
+  TransactionBatchRequest,
+  TransactionController,
+  TransactionControllerMessenger,
+} from '..';
 import { projectLogger } from '../logger';
 import {
   TransactionEnvelopeType,
-  type TransactionBatchRequest,
   type TransactionBatchResult,
   type TransactionParams,
   TransactionType,
@@ -29,6 +34,7 @@ type AddTransactionBatchRequest = {
   getEthQuery: (networkClientId: string) => EthQuery;
   getInternalAccounts: () => Hex[];
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
   request: TransactionBatchRequest;
 };
 
@@ -36,6 +42,7 @@ type IsAtomicBatchSupportedRequest = {
   address: Hex;
   getEthQuery: (chainId: Hex) => EthQuery;
   messenger: TransactionControllerMessenger;
+  publicKeyEIP7702?: Hex;
 };
 
 const log = createModuleLogger(projectLogger, 'batch');
@@ -54,15 +61,25 @@ export async function addTransactionBatch(
     getChainId,
     getInternalAccounts,
     messenger,
+    publicKeyEIP7702,
     request: userRequest,
   } = request;
+
+  const sizeLimit = getBatchSizeLimit(messenger);
 
   validateBatchRequest({
     internalAccounts: getInternalAccounts(),
     request: userRequest,
+    sizeLimit,
   });
 
-  const { from, networkClientId, requireApproval, transactions } = userRequest;
+  const {
+    batchId: batchIdOverride,
+    from,
+    networkClientId,
+    requireApproval,
+    transactions,
+  } = userRequest;
 
   log('Adding', userRequest);
 
@@ -75,9 +92,14 @@ export async function addTransactionBatch(
     throw rpcErrors.internal('Chain does not support EIP-7702');
   }
 
+  if (!publicKeyEIP7702) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
+
   const { delegationAddress, isSupported } = await isAccountUpgradedToEIP7702(
     from,
     chainId,
+    publicKeyEIP7702,
     messenger,
     ethQuery,
   );
@@ -101,6 +123,7 @@ export async function addTransactionBatch(
     const upgradeContractAddress = getEIP7702UpgradeContractAddress(
       chainId,
       messenger,
+      publicKeyEIP7702,
     );
 
     if (!upgradeContractAddress) {
@@ -113,14 +136,15 @@ export async function addTransactionBatch(
 
   log('Adding batch transaction', txParams, networkClientId);
 
-  const { transactionMeta, result } = await addTransaction(txParams, {
+  const batchId = batchIdOverride ?? generateBatchId();
+
+  const { result } = await addTransaction(txParams, {
+    batchId,
     nestedTransactions,
     networkClientId,
     requireApproval,
     type: TransactionType.batch,
   });
-
-  const batchId = transactionMeta.id;
 
   // Wait for the transaction to be published.
   await result;
@@ -139,7 +163,16 @@ export async function addTransactionBatch(
 export async function isAtomicBatchSupported(
   request: IsAtomicBatchSupportedRequest,
 ): Promise<Hex[]> {
-  const { address, getEthQuery, messenger } = request;
+  const {
+    address,
+    getEthQuery,
+    messenger,
+    publicKeyEIP7702: publicKey,
+  } = request;
+
+  if (!publicKey) {
+    throw rpcErrors.internal('EIP-7702 public key not specified');
+  }
 
   const chainIds7702 = getEIP7702SupportedChains(messenger);
   const chainIds: Hex[] = [];
@@ -150,6 +183,7 @@ export async function isAtomicBatchSupported(
     const { isSupported, delegationAddress } = await isAccountUpgradedToEIP7702(
       address,
       chainId,
+      publicKey,
       messenger,
       ethQuery,
     );
@@ -162,4 +196,15 @@ export async function isAtomicBatchSupported(
   log('Atomic batch supported chains', chainIds);
 
   return chainIds;
+}
+
+/**
+ * Generate a tranasction batch ID.
+ *
+ * @returns  A unique batch ID as a hexadecimal string.
+ */
+function generateBatchId(): Hex {
+  const idString = v4();
+  const idBytes = new Uint8Array(parse(idString));
+  return bytesToHex(idBytes);
 }